pyproject.toml
[project]
name = "bank-course-work-1"
version = "0.1.0"
description = "Домашнее задание по анализу банковских данных"
readme = "README.md"
authors = [
    {name = "Ангелина Павленко", email = "angel-ina@mail.ru"}
]
requires-python = ">=3.11"
dependencies = [
    "pandas>=3.0.1,<4.0.0",
    "openpyxl>=3.1.5,<4.0.0",
    "requests>=2.32.5,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2,<10.0.0",
    "flake8>=7.3.0,<8.0.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

